requirements.txt
aioredis==1.3.1
amqp==5.0.2
anyjson==0.3.3
asgiref==3.3.1
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
bcrypt==3.2.0
billiard==3.6.3.0
cached-property==1.5.2
celery==5.0.3
certifi==2020.11.8
cffi==1.14.4
channels==3.0.2
channels-redis==3.2.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
configparser==5.0.1
constantly==15.1.0
cryptography==3.2.1
daphne==3.0.1
distro==1.5.0
dj-database-url==0.5.0
Django==3.1.7
django-braces==1.14.0
django-daterange-filter==1.3.0
django-filter==2.4.0
django-jsonfield==1.4.1
django-model-utils==4.1.1
django-redis==4.12.1
django-simple-templates==0.7.1
django-timezone-field==4.1.1
djangorestframework==3.12.2
docopt==0.6.2
flake8==3.9.2
gunicorn==20.0.4
hiredis==1.1.0
httplib2==0.18.1
hyperlink==20.0.1
idna==2.10
importlib-metadata==3.1.1
incremental==17.5.0
isort==5.7.0
Jinja2==2.11.2
jsonschema==3.2.0
kombu==5.0.2
lazy-object-proxy==1.4.3
Markdown==3.3.3
mccabe==0.6.1
msgpack==1.0.0
msgpack-python==0.5.6
packaging==20.7
paramiko==2.7.2
parsedatetime==2.6
pathlib==1.0.1
prompt-toolkit==3.0.8
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyHamcrest==2.0.2
PyNaCl==1.4.0
pyOpenSSL==20.0.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.3
python-dotenv==0.15.0
pytz==2020.4
PyYAML==5.3.1
redis==3.5.3
requests==2.25.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
texttable==1.6.3
toml==0.10.2
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.2
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
websocket-client==0.57.0
whitenoise==5.2.0
wrapt==1.12.1
zipp==3.4.0
zope.interface==5.2.0
# whitenoise==5.2.0
